requirements.txt
streamlit==1.44.0

google.generativeai== 0.8.4
PyPDF2==3.0.1